Add maxPassengers field and Motorcycle class
The maxPassengers field has been added to the Vehicle class to indicate the maximum passenger capacity. The Motorcycle class, a new subclass of Vehicle, has also been created. This includes relevant fields and a toString method for output purposes. Subsequently, an instance of Motorcycle is created and used in the Main class.